# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface DashboardSummary {
  totalSales: number;
  totalItemsSold: number;
  totalRegions: number;
  totalShops: number;
}

export interface SalesTimePoint {
  month: string;
  sales: number;
}

export interface SalesByShop {
  shopName: string;
  shopId: number;
  sales: number;
}

export interface SalesByRegion {
  regionName: string;
  regionId: number;
  sales: number;
}

export interface Region {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  shopCount: number;
}

export interface CreateRegionBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateRegionBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface Shop {
  id: number;
  regionId: number;
  name: string;
  /** @nullable */
  address: string | null;
  /** @nullable */
  contactPhone: string | null;
  createdAt: string;
}

export interface CreateShopBody {
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactPhone?: string | null;
}

export interface UpdateShopBody {
  name?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactPhone?: string | null;
}

export interface Item {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  unitPrice: number;
  minOrderQty: number;
  stockQuantity: number;
  /** @nullable */
  category: string | null;
  /** @nullable */
  sku: string | null;
  createdAt: string;
}

export interface CreateItemBody {
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  minOrderQty: number;
  stockQuantity: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
}

export interface UpdateItemBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  unitPrice?: number;
  minOrderQty?: number;
  stockQuantity?: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
}

export interface Order {
  id: number;
  shopId: number;
  totalAmount: number;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  itemCount: number;
}

export interface OrderLineItem {
  id: number;
  itemId: number;
  itemName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export interface OrderDetail {
  id: number;
  shopId: number;
  totalAmount: number;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  lineItems: OrderLineItem[];
}

export interface CreateOrderLineItem {
  itemId: number;
  quantity: number;
}

export interface CreateOrderBody {
  /** @nullable */
  notes?: string | null;
  lineItems: CreateOrderLineItem[];
}

export type GetRegionSalesDetailParams = {
  regionId: number;
};