# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inventory & Sales Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /regions:
    get:
      operationId: getRegions
      tags: [regions]
      summary: List all regions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Region"
    post:
      operationId: createRegion
      tags: [regions]
      summary: Create a new region
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegionBody"
      responses:
        "201":
          description: Region created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
  /regions/{id}:
    get:
      operationId: getRegion
      tags: [regions]
      summary: Get a region by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "404":
          description: Region not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRegion
      tags: [regions]
      summary: Update a region
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegionBody"
      responses:
        "200":
          description: Region updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "404":
          description: Region not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRegion
      tags: [regions]
      summary: Delete a region
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Region deleted
        "404":
          description: Region not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /items:
    get:
      operationId: getItems
      tags: [items]
      summary: List all inventory items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create an inventory item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    patch:
      operationId: updateItem
      tags: [items]
      summary: Update an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an inventory item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Item deleted
  /shops/{shopId}/orders:
    get:
      operationId: getOrdersByShop
      tags: [orders]
      summary: Get orders for a shop
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order for a shop
      security:
        - bearerAuth: []
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Order deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - createdAt
...
    Region:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
        shopCount:
          type: integer
      required:
        - id
        - name
        - description
        - createdAt
        - shopCount
    CreateRegionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
    UpdateRegionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
...
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        minOrderQty:
          type: integer
        stockQuantity:
          type: integer
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - unitPrice
        - minOrderQty
        - stockQuantity
        - category
        - sku
        - createdAt
    CreateItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        minOrderQty:
          type: integer
        stockQuantity:
          type: integer
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
      required:
        - name
        - unitPrice
        - minOrderQty
        - stockQuantity
    UpdateItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        unitPrice:
          type: number
        minOrderQty:
          type: integer
        stockQuantity:
          type: integer
        category:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
    Order:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        itemCount:
          type: integer
      required:
        - id
        - shopId
        - totalAmount
        - status
        - notes
        - createdAt
        - itemCount
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
      required:
        - id
        - shopId
        - totalAmount
        - status
        - notes
        - createdAt
        - lineItems
    OrderLineItem:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required:
        - id
        - itemId
        - itemName
        - quantity
        - unitPrice
        - subtotal
    CreateOrderBody:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderLineItem"
      required:
        - lineItems
    CreateOrderLineItem:
      type: object
      properties:
        itemId:
          type: integer
        quantity:
          type: integer
      required:
        - itemId
        - quantity