# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateItemBody {
  name: string;
  /** @nullable */
  description?: string | null;
  unitPrice: number;
  minOrderQty: number;
  stockQuantity: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
}