# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderLineItem } from "./createOrderLineItem";

export interface CreateOrderBody {
  /** @nullable */
  notes?: string | null;
  lineItems: CreateOrderLineItem[];
}