# lib/api-zod/src/generated/types/createRegionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRegionBody {
  name: string;
  /** @nullable */
  description?: string | null;
}