# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  unitPrice: number;
  minOrderQty: number;
  stockQuantity: number;
  /** @nullable */
  category: string | null;
  /** @nullable */
  sku: string | null;
  createdAt: string;
}