# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  shopId: number;
  totalAmount: number;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  itemCount: number;
}