# lib/api-zod/src/generated/types/orderDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory & Sales Management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineItem } from "./orderLineItem";

export interface OrderDetail {
  id: number;
  shopId: number;
  totalAmount: number;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  lineItems: OrderLineItem[];
}